projects/as-harlamov/lab2/source/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Document parsing
pypdf>=3.17.0
python-docx>=1.1.0
python-pptx>=0.6.23
markdownify>=0.11.6
beautifulsoup4>=4.12.0

# Text processing and chunking
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
tiktoken>=0.5.0

# Embeddings
sentence-transformers>=2.2.2
openai>=1.10.0
rank-bm25>=0.2.2

# Vector store
qdrant-client>=1.7.0

# LLM integration
openai>=1.10.0
anthropic>=0.18.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0

